chore: check package and cli command wip
