+++ packages/cli/package.json
@@ -46,6 +46,7 @@
     "@hono/node-server": "^1.11.0",
     "@parcel/watcher": "^2.4.1",
     "@scalar/api-reference": "workspace:*",
+    "@scalar/config": "workspace:*",
     "@scalar/mock-server": "workspace:*",
     "@scalar/oas-utils": "workspace:*",
     "@scalar/openapi-parser": "^0.7.2",
