+++ packages/cli/src/commands/check/invalid.json
@@ -1,0 +1,7 @@
+{
+  "myobj": {
+    "path": "src/page-one.md",
+    "type": "page",
+    "name": "My First Page!!"
+  }
+}
