+++ packages/cli/src/commands/check/valid.json
@@ -28,7 +28,7 @@
   ],
   "references": [],
   "siteConfig": {
-    "footer": "<footer>\r\n  <p>This is the testing footer!!</p>\r\n</footer>",
-    "theme": "some-theme"
+    "theme": "some-theme",
+    "footer": "<footer>\r\n  <p>This is the testing footer!!</p>\r\n</footer>"
   }
 }
