+++ packages/cli/tsconfig.json
@@ -1,5 +1,4 @@
 {
   "extends": "../../tsconfig.json",
-  "compilerOptions": {},
-  "include": ["src", "rollup.config.ts", "../config/src/check.ts"]
+  "compilerOptions": {}
 }
