+++ packages/config/schemas/config.json
