+++ packages/config/test/check.test.ts
@@ -1,7 +1,6 @@
-import path from 'node:path'
 import { describe, expect, it } from 'vitest'
 
-import { check } from '../src/check'
+import { check } from '../src/index'
 
 describe('check', () => {
   it('checks the valid json file', () => {
@@ -12,6 +11,6 @@ describe('check', () => {
   it('checks the invalid json file', () => {
     const result = check('/test/invalid.json')
     console.log(JSON.stringify(result, null, 3))
-    expect(result.valid).toBe(true)
+    expect(result.valid).toBe(false)
   })
 })
